Add "targets" docs to dev server
There are four targets that DCAR can render content for, and this change
adds documentation pages for each of them, on the following paths:

- `/targets/dotcom`
- `/targets/live-apps`
- `/targets/editions-app`
- `/targets/amp`

The links on the `/targets` page now work. However the links on each of
these new documentation pages do not yet point anywhere. These pages
will be added in a future update, in the interests of keeping changes
smaller and easier to review.