`Versus` election trackers component
Adds a new election trackers component to display two groups facing one
another in an election. This could be used, for example, to show two
parties attempting to achieve a majority in Parliament. The included
stories show examples of this.

This component is one of several that will allow rendering election
trackers within DCAR.