Add "pages" And "targets" To Dev Server (#13868)
This begins adding some documentation to the DCAR dev server, to provide
more information about what the project can do. There are two ways of
breaking this down. You can either describe the project in terms of the
targets it can render content for, such as dotcom and the live apps, or
you can describe it in terms of the pages it can render, like articles,
fronts, and so on.

This change provides documentation structured in both ways, as either
can be useful for understanding the project. It introduces two new
paths:

- `/pages`
- `/targets`

each with corresponding documentation and links to further information.
The links do not yet point anywhere as their content will be added in
subsequent updates, to keep changes smaller and easier to review.