Add galleries to the article model
The article design, display and theme have been moved to the root of the
article model in order to use the design as a discriminated field for
different article types. This way, different articles can have different
fields from one another. In this case, galleries can have the list of
images which is all they contain.

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>