`ValuesWithChange` Election Trackers Component (#13920)
Adds a new election trackers component to display a list of values and
the amount by which they've changed. This could be used, for example, to
display a list of seats won in an election along with the changes since
the last election. The included stories show examples of this.

This component is one of several that will allow rendering election
trackers within DCAR.