Add the rendering set up for the HideMobileHighlights test (#14008)
* Adds the rendering set up for the HideMobileHighlights test.

In the test, if the user is in the *variant*, then we hide the highlights carousel until the tablet breakpoint. Otherwise, if the user is in the control or otherwise, we render the highlights carousel at all breakpoints.

The test is running 5% of users and is running for 6 days. After which point this PR can be reverted.

* Only hide the highlights container on the uk front