Refactor dateline text colour
There are a few issues with the dateline text colour, including it
appearing:

- Too close to the background colour on picture, video and audio
articles at narrower breakpoints.
- Inconsistent across different breakpoints within the same article.
- Inconsistent between apps and web.

This change fixes some bugs and attempts to flatten out many of the
inconsistencies. Some of the consequences of this change include:

- The colour on picture, audio and video articles is now much lighter,
making it clearly stand out from the background.
- The colour is now typically the same across different breakpoints
within the same article, except in the case of liveblogs where the
design is necessarily different at different breakpoints.
- The text is now darker on a number of web articles (e.g. features),
matching the design on apps.
- The text is now darker on some apps articles (e.g. comment), matching
the design on web.