Update live block datetime colour
The blocks on liveblogs (and deadblogs) have up to two possible
datetimes. They have a "first published" datetime that appears at the
top and, optionally, a "last updated" datetime that appears in the
bottom right.

The colours for these did not look correct in dark mode because they
only had hardcoded light mode versions. This change moves these colours
into the palette and introduces dark mode versions. It also updates the
colour from the `neutral` palette used for light mode.

In addition, previously the hover colour was relying on a `filter`,
which did not work correctly in dark mode. This change replaces this
with a dedicated hover colour in the palette.