Fix `Toast` stories
Issues in the type definition of the `meta` object meant missing `args`
in the stories were not being reported correctly by TypeScript. An
upgrade of the storybook packages surfaced this problem through errors
reported on the `Template` function, which was being used as a custom
`render` function for the stories.

Fixing the types on `meta`, adding the missing `args`, and inlining the
`render` function to make use of automatic type inference fixes this
issue.