Add Cricket Docs To Dev Server (#14097)
Adds documentation for the cricket scorecard page, which can be rendered
on dotcom only.

The cricket scorecard pages go out of date over time and no longer
provide cricket data. In addition, the scorecards that appear on
liveblogs also go out of date and disappear from the page. Therefore
we should not link directly to these pages from the docs, but instead
link to a list of recent cricket liveblogs. These should have up-to-date
data, and links to the scorecard pages.