Add A Nav For Football Competitions (#14170)
When major competitions are running we'd like to introduce additional
navigation to the football data pages that relate to that competition:
fixtures, live matches, results and tables. This will allow users to
navigate between different pieces of information about said competition.

This change introduces a `FootballCompetitionNav` component to be
displayed only when the `pageId` matches the given competition. For now
this is hardcoded to the Women's Euro 2025 competition as a first
implementation. We will make it generic and configurable for other
competitions in a later change.