Remove `data-print-layout` from `Section`
If the `data-print-layout` attribute is applied to an element and set to
`'hide'`, the `print.css` stylesheet is used to target some styles that
hide the element. However, this only works if the attribute is applied
to an element in the DOM. If the attribute is instead applied as a prop
to a component, and that component does not expect that prop and apply
it in turn to an actual DOM element, then it's simply ignored.

The `Section` component does not take this prop, therefore these usages
of `data-print-layout` have no effect.