Update Node to v22.18.0
The latest version of v22 (Active LTS) at the time of this change. This
keeps Node up to date.

AR also had an out-of-date version of `types/node`, which is now updated
to 22.x in this change.

Also updates webpack to latest to fix a TS error in its type
declarations surfaced by these updates.