Add a new homepage to the dev server
It contains a form to input a dotcom URL. Submitting the form will query
frontend for information about that page, and then present a summary of
this to the user. Included in that summary are links for opening the
dotcom and apps versions of that page locally, in CODE, and on PROD.

This supports only a single article URL for now, and the summary
provides information on the article format, its main media, and the body
elements it contains. Support for more kinds of pages (fronts for
example), and viewing information about multiple pages at once, will
come in a future update.

As the new page does not yet have feature parity with the old homepage,
it does not replace it for now, and is instead available on `/home`.