Adds an media field to the data link name array.
This field indicates what media type the card contains. If there is no media, this is resolved to `none`.

We set the link name during card enhancement but we do not know at this stage what the card's media is. This is because we do not know details such as card size, which can impact rendered media.

As a result of this, we potentially override the media section of the link name at the card level IF media is displayed on the card.