Track card media in data-link-name  (#14400)
* Rename video -> youtube-video as this is only used for youtube videos.

This makes tracking the video context somewhat easier

* Adds an media field to the data link name array.

This field indicates what media type the card contains. If there is no media, this is resolved to `none`.

We set the link name during card enhancement but we do not know at this stage what the card's media is. This is because we do not know details such as card size, which can impact rendered media.

As a result of this, we potentially override the media section of the link name at the card level IF media is displayed on the card.

* Update naming convention of video -> youtube-video

* Rename CardImageType to CardMediaType and use shared type in the linkName setter

* Prefer an appender function to overwriting the media type