Add story packages to galleries
The primary reason for this change is to add support for story packages
to the new gallery layout.

At the moment every layout duplicates the same transformation on the
frontend story package data, mapping over the trails with `decideTrail`.
This change extracts the story package type definition and parsing into
its own module, and calls the parsing code up front when the `Article`
type is constructed (i.e. in one place rather than several). For now, we
can only use this in galleries, as other layouts are still using the
`ArticleDeprecated` type.

Also adds a type annotation to the story package fixtures, so that we
can use them in stories without type errors.