Remove `z-index` from gallery main media (#14599)
It isn't needed for the layout. The lightbox link is absolute positioned
relative to the figure, which is itself relative positioned; this allows
the link to precisely overlay the image.

Relative positioning the figure, however, causes it to be drawn over the
headline and series link. These elements would normally appear over the
top of the image because they are a) placed in the same location
with CSS grid, and b) later in the DOM. Relatively positioning these
elements solves this problem, and allows them to overlay the image
again.

In summary, this layout has already been achieved by positioning the
elements, so setting additional `z-index` properties is not required.