`LeftColumn` should not use `section` (#14611)
The `section` element is meant to be used for a standalone section of a
document. This generally is not what this `LeftColumn` component is used
to represent. For example, it's most widely used to position a heading
within the `Section` component, which typically already includes a
`section` itself.

`LeftColumn` is for general-purpose, presentational usage, and therefore
can't make assumptions about the semantics of the document in which it's
used. As a result, a non-semantic element like `div` is a more
appropriate choice.