Prop drill server time
The `serverTime` prop replaces the `absoluteServerTimes` boolean and is
intended to allow us to fetch the server time once at the top level and
pass it down to child components that need it. Currently `DateTime`
calls `Date.now()`, but this may be called on the server and client
leading to hydration errors where the time has drifted.