Gallery Sub-Components: `Body` (#14626)
Refactoring the `GalleryLayout` into sub-components, to make it easier
to read.

ESLint reports that a function doesn't have a `return` statement, but
it does. It's possible ESLint thinks the switch might not be exhaustive,
but it is and TypeScript is able to verify this. Therefore it may be an
issue with the integration between TS and ESLint.