Remove 'true anonymity'
Discussed with Zoe Norden who didn't want to suggest that anonymity was guaranteed in perpetuity: only the Secure Messaging service has sufficient protections for full anonymity and the lifecycle of a correspondence may extend beyond that. Because there isn't space to explain all that we opted to just refer to "alternative methods" without making promises that go beyond technology.