Follow `no-unused-expressions` lint rule (#14691)
This rule comes from `typescript-eslint`:
https://typescript-eslint.io/rules/no-unused-expressions/

We don't have it enabled yet, but in v8 it's part of the "recommended"
preset:
https://typescript-eslint.io/blog/announcing-typescript-eslint-v8/#updated-configuration-rules

In order to keep the upgrade to that version as small as possible, this
change pre-emptively fixes code considered incorrect by that rule. A
description of these fixes follows.

Replaced `&&` expressions where the result isn't used, with conditional
statements using `if`:

```ts
a && sideEffect(a);
```

becomes:

```ts
if (a) {
    sideEffect(a);
}
```

Replaced ternary expressions where the result isn't used with
conditional statements using `if...else`:

```ts
a ? sideEffect(a) : otherEffect(a);
```

becomes:

```ts
if (a) {
    sideEffect(a);
} else {
    otherEffect(a);
}
```

Replaced some usages of the comma operator with semi-colon. The comma
operator evaluates all of its operands, returns the value of the last
operand, and discards all others. In the cases updated here, the results
of the expressions weren't used, so these operands are better written as
statements, i.e. separated with a semi-colon:

```ts
sideEffect(a), otherEffect(a);
```

becomes:

```ts
sideEffect(a); otherEffect(a);
```

See also:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Comma_operator

Ignored the case in `Island.test.tsx` where unused expressions are
included intentionally to test compilation. Also removed some
expressions that weren't doing anything, and may therefore have been
left behind in error.