Follow `no-empty-object-type` lint rule (#14694)
This rule comes from `typescript-eslint`:
https://typescript-eslint.io/rules/no-empty-object-type/

We don't have it enabled yet, but in v8 it's part of the "recommended"
preset:
https://typescript-eslint.io/blog/announcing-typescript-eslint-v8/#updated-configuration-rules

In order to keep the upgrade to that version as small as possible, this
change pre-emptively fixes code considered incorrect by that rule. The
issues all relate to empty interfaces, for which the linter points out
that:

```
An interface declaring no members is equivalent to its supertype
```

Solutions to this include using the supertype directly, or type aliasing
the supertype to the new name. This change opts for the latter:

```ts
interface SuperType { a: string };

interface SubType extends SuperType {};
```

becomes:

```ts
interface SuperType { a: string };

type SubType = SuperType;
```

There are some differences between type aliases and interfaces:
https://www.typescriptlang.org/docs/handbook/2/everyday-types.html#differences-between-type-aliases-and-interfaces

If we prefer to keep the interfaces, we can instead customise this lint
rule in DCAR, setting `allowInterfaces` to `'with-single-extends'`:
https://typescript-eslint.io/rules/no-empty-object-type/#allowinterfaces

The cost would be diverging from our centralised linting rules, and
therefore adding some complexity to DCAR.