Follow `prefer-promise-reject-errors` lint rule (#14693)
This rule comes from `typescript-eslint`:
https://typescript-eslint.io/rules/prefer-promise-reject-errors/

We don't have it enabled yet, but in v8 it's part of the
"recommended-type-checked" preset:
https://typescript-eslint.io/blog/announcing-typescript-eslint-v8/#updated-configuration-rules

In order to keep the upgrade to that version as small as possible, this
change pre-emptively fixes code considered incorrect by that rule. It
creates `Error` objects for instances where we have control over the
value the `Promise` is rejecting with, and disables the rule in cases
where we don't have control over this value because it comes from a
library.

Another approach would be to create our own `Error` object and include
the unknown library error using the `cause` option, but the browser
support is not good enough for that yet:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error/Error#cause