Follow `no-unused-expressions` lint rule in AR (#14748)
This rule comes from `typescript-eslint`:
https://typescript-eslint.io/rules/no-unused-expressions/

We don't have it enabled yet, but in v8 it's part of the "recommended"
preset:
https://typescript-eslint.io/blog/announcing-typescript-eslint-v8/#updated-configuration-rules

In order to keep the upgrade to that version as small as possible, this
change pre-emptively fixes code considered incorrect by that rule. A
description of these fixes follows.

Replaced `&&` expressions where the result isn't used, with conditional
statements using `if`:

```ts
a && sideEffect(a);
```

becomes:

```ts
if (a) {
    sideEffect(a);
}
```