Remove AR CI checks
We're removing the project, and therefore no longer need these to run.
`ar-nodejs` runs the validation checks on AR code. The labelling
workflow is used to label PRs as either "apps-rendering" or
"dotcom-rendering". With only "dotcom-rendering" left we no longer need
this.

Note that two workflows related to AR remain:

- `ar-chromatic` - will be removed in a separate change that involves
removing the Chromatic integration.
- `ar-ci` - will be removed in a separate change that involves removing
the AR infra.