Extend expiry of user benefits cookies to 30 days
Previouly they were short lived (1-2 days) but this resulted in edge
cases where if a signed in user didn't visit the site for more than a
couple of days, when they returned their first page view wouldn't reflect
their benefits (i.e. they would see ads). This is due to a race
condition between the user benefits refresh and the ads code. However,
we don't want to delay ads until after the user benefits have been
refreshed as that would impact performance. So instead, extend the
expiry of the cookie.

Note: this may result in a user getting benefits they no longer have on
the first returning pageview, but this will be correct from the second
page view onwards. We think this is OK.

This also means we now remove cookies if the user benefits response says
they no longer have the benefit, rather than simply letting the cookie
expire.