Don't render `EditDialog` when not editing
Fixes a bug where the state for the form fields wasn't being populated
by the initial values coming from props.

As the form was always rendered, regardless of whether anything was
currently being edited, the initial values coming from the props and
passed to the `useState` calls were empty. Then when a particular
section was selected, these props changed to reflect the data from that
section. However, `useState` only uses the initial values from the
*first render*, so it was ignoring these updated props, and retaining
the empty values.

The solution is not to render the form while nothing is being edited,
and then render it when a section is selected for editing. That way each
time a section is selected for editing, the form is considered a new
addition to the tree, and therefore the initial data coming from props
on the "first render" is the section data.