Don't render `InsertDialog` when not editing
Fixes a bug where the state for the form fields was being populated by
the data from the previous section added.

As the form is always rendered, regardless of whether anything is
currently being inserted, the state shown in the form fields and managed
by `useState` persists for the lifetime of the application. The first
time someone adds a section, the fields are empty. But if they're filled
out, the section inserted, and then the form is opened again to insert
another section, it still retains the values from the previous section
added.

The solution is not to render the form while nothing is being inserted,
and then render it when the "Add Section" button is clicked. That way
each time the form appears it is considered a new addition to the tree,
and therefore the initial, empty data coming from props on the "first
render" is used to populate the state.