Image resizer for football crests (#14939)
Currently for football crests we are limited to two prebuilt sizes of
image, 60px wide and 120px wide, hosted on `sport.guim.co.uk`. We would
instead like to use the flexibility of the image resizer, on
`i.guim.co.uk`, to request a greater range of image variants. This will
allow us to choose more appropriate image dimensions for different
designs, give us access to more image formats, and allow us to provide
responsive images for different screen pixel densities.

This PR also refactors several different implementations of the crest
image element into a pair of components, `FootballCrest` and
`FootballTableCrest`, and consolidates the building of the crest URL
within one of these.