Delete `apps-rendering` directory (#14966)
We're turning off this project and deleting the code. Note that this
change focuses on deleting the directory, and only makes changes to code
outside it that would otherwise break with the directory gone. This is
to keep this change as small as possible. Further references will be
deleted in later changes.