New football models (#14995)
A refactoring of our existing football models, to reduce duplication and
allow them to work for upcoming changes to football sportblogs and match
report pages.

These are not currently being used anywhere. The work to have them
replace the existing models, and be used on the aforementioned pages,
will happen in future changes.