Football match header
The new designs for the football match header that will appear across
match info pages, liveblogs, and match reports. It makes use of a
similar implementation for displaying scores as the previous version,
i.e. SVGs, but uses the existing `BigNumber` component.

Note that this does not yet include the designs for the live version.